+++ src/test/java/com/truward/semantic/id/test/IdCodecTest.java
@@ -5,44 +5,111 @@ import com.truward.semantic.id.SemanticIdCodec;
 import com.truward.semantic.id.exception.IdParsingException;
 import org.junit.Test;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Random;
+import java.util.*;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
+import static org.junit.Assert.*;
 
 /**
  * @author Alexander Shabanov
  */
-public class IdCodecTest {
+public final class IdCodecTest {
 
   @Test
   public void shouldSetNames() {
-    final SemanticIdCodec codec = SemanticIdCodec.forService("foo1").withEntityName("user");
-    assertEquals("foo1", codec.getServiceName());
+    final SemanticIdCodec codec = SemanticIdCodec.forService("foo2").withEntityName("user");
+    assertEquals("foo2", codec.getServiceName());
     assertEquals("user", codec.getEntityName());
+  }
+
+  @Test
+  public void shouldBeAbleToDecodeIdWithServiceName() {
+    final SemanticIdCodec codec1 = SemanticIdCodec.forService("f1");
+    final SemanticIdCodec codec2 = SemanticIdCodec.forService("f2");
+
+    assertTrue(codec1.canDecode("f1.abc"));
+    assertTrue(codec2.canDecode("f2.abc"));
+
+    assertFalse(codec1.canDecode("f2.abc"));
+    assertFalse(codec2.canDecode("f1.abc"));
+
+    assertFalse(codec1.canDecode("foo1.user.abc"));
+    assertFalse(codec2.canDecode("foo2.user.abc"));
+    assertFalse(codec1.canDecode("foo2.user.abc"));
+    assertFalse(codec2.canDecode("foo1.user.abc"));
+  }
+
+  @Test
+  public void shouldBeAbleToDecodeIdWithServiceNameAndEntityName() {
+    final SemanticIdCodec codec1 = SemanticIdCodec.forService("foo1").withEntityName("user");
+    final SemanticIdCodec codec2 = SemanticIdCodec.forService("foo2").withEntityName("user");
+
+    assertTrue(codec1.canDecode("foo1.user.abc"));
+    assertTrue(codec2.canDecode("foo2.user.abc"));
+
+    assertFalse(codec1.canDecode("f1.abc"));
+    assertFalse(codec2.canDecode("f2.abc"));
+    assertFalse(codec1.canDecode("foo2.user.abc"));
+    assertFalse(codec2.canDecode("foo1.user.abc"));
   }
 
   @Test
   public void shouldEncodeAndDecodeIdWithSingleServiceName() {
     final IdCodec codec = SemanticIdCodec.forService("foo1");
     assertEquals("foo1.1", codec.encodeLong(1));
+    assertTrue(codec.canDecode("foo1.1"));
+    assertTrue(codec.canDecode("foo1.2"));
     assertEquals(1, codec.decodeLong("foo1.1"));
-    checkTestLongs(codec);
+    assertFalse(codec.canDecode("foo2.1"));
+    checkTestValues(codec);
     assertDecodeLongFails(codec, Arrays.asList("foo1.", "foo1.-", "foo1.012345678901234", "foo2.1"));
   }
 
   @Test
   public void shouldEncodeAndDecodeIdWithServiceAndEntityName() {
-    final IdCodec codec = SemanticIdCodec.forService("foo1").withEntityName("user");
-    assertEquals("foo1.user.1", codec.encodeLong(1));
-    assertEquals(1, codec.decodeLong("foo1.user.1"));
-    checkTestLongs(codec);
-    assertDecodeLongFails(codec, Arrays.asList("foo1.user.", "foo1.user.-", "foo1.user.12345678901234",
-        "foo2.user.1", "foo1.book.1"));
+    final IdCodec codec = SemanticIdCodec.forService("foo1").withEntityName("account");
+    assertEquals("foo1.account.1", codec.encodeLong(1));
+    assertEquals(1, codec.decodeLong("foo1.account.1"));
+    checkTestValues(codec);
+    assertDecodeLongFails(codec, Arrays.asList("foo1.account.", "foo1.account.-", "foo1.account.12345678901234",
+        "foo2.account.1", "foo1.book.1"));
+  }
+
+  @Test
+  public void shouldEncodeVariadicLengthByteArrayIds() {
+    final IdCodec codec = SemanticIdCodec.forService("book1");
+    final List<byte[]> ids = Arrays.asList(
+        new byte[] { 0 },
+        new byte[] { 1 },
+        new byte[] { -1 },
+        new byte[] { Byte.MAX_VALUE },
+        new byte[] { Byte.MIN_VALUE },
+        new byte[] { 0, 0 },
+        new byte[] { 1, Byte.MAX_VALUE },
+        new byte[] { Byte.MIN_VALUE, -1 },
+        new byte[] { Byte.MIN_VALUE, Byte.MIN_VALUE, Byte.MIN_VALUE },
+        new byte[] { Byte.MAX_VALUE, Byte.MAX_VALUE, Byte.MAX_VALUE },
+        new byte[] { 1, 1, 1, 1, 1 },
+        new byte[] { 1, -1, 127, 63, -67 }
+    );
+
+    for (final byte[] id : ids) {
+      final String strId = codec.encodeBytes(id);
+      assertArrayEquals(id, codec.decodeBytes(strId));
+    }
+
+    // check encode/decode of the other arrays
+    final Random random = new Random(34287562345L);
+    for (int size = 1; size < IdCodec.MAX_BYTES_ID_SIZE; ++size) {
+      final byte[] id = new byte[size];
+      for (int tries = 0; tries < 10; ++tries) {
+        for (int i = 0; i < size; ++i) {
+          random.nextBytes(id);
+        }
+
+        final String strId = codec.encodeBytes(id);
+        assertArrayEquals(id, codec.decodeBytes(strId));
+      }
+    }
   }
 
   private static void assertDecodeLongFails(IdCodec codec, List<String> malformedIds) {
@@ -56,7 +123,7 @@ public class IdCodecTest {
     }
   }
 
-  private static void checkTestLongs(IdCodec codec) {
+  private static void checkTestValues(IdCodec codec) {
     final List<Long> testLongs = new ArrayList<>();
     testLongs.addAll(Arrays.asList(0L, 1L, 1000L, Long.MAX_VALUE, Long.MIN_VALUE, -1L, -1000L));
     final Random random = new Random(1248946974651564L);
@@ -69,6 +136,23 @@ public class IdCodecTest {
       final long otherVal = codec.decodeLong(semanticId);
       final long otherVal1 = codec.decodeLong(semanticId.toUpperCase());
       final long otherVal2 = codec.decodeLong(semanticId.toLowerCase());
+
+      assertEquals("Converted value does not match original one for " + codec, val, otherVal);
+      assertEquals("[Uppercase] Converted value does not match original one for " + codec, val, otherVal1);
+      assertEquals("[Lowercase] Converted value does not match original one for " + codec, val, otherVal2);
+    }
+
+    final List<UUID> testUuids = new ArrayList<>();
+    testUuids.addAll(Arrays.asList(new UUID(0L, 0L), new UUID(Long.MAX_VALUE, Long.MAX_VALUE)));
+    for (int i = 0; i < 1000; ++i) {
+      testUuids.add(new UUID(random.nextLong(), random.nextLong()));
+    }
+
+    for (final UUID val : testUuids) {
+      final String semanticId = codec.encodeUUID(val);
+      final UUID otherVal = codec.decodeUUID(semanticId);
+      final UUID otherVal1 = codec.decodeUUID(semanticId.toUpperCase());
+      final UUID otherVal2 = codec.decodeUUID(semanticId.toLowerCase());
 
       assertEquals("Converted value does not match original one for " + codec, val, otherVal);
       assertEquals("[Uppercase] Converted value does not match original one for " + codec, val, otherVal1);