+++ RELEASE-NOTES.md
@@ -1,6 +1,11 @@
 Release Notes: semantic-id
 ==========================
 
+# 1.0.1
+
+* Add encode and decode methods for byte array and UUID keys
+* Add canDecode method
+
 # 1.0.0
 
 Initial version